Suggestion: Extract annotation logic from Router into factory
